refactor remote data source
