+++ data/build.gradle.kts
@@ -1,3 +1,5 @@
+import java.util.Properties
+
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
@@ -14,9 +16,14 @@ android {
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles("consumer-rules.pro")
     }
-
+    android.buildFeatures.buildConfig = true
+    val properties = Properties()
+    properties.load(rootProject.file("secret.properties").inputStream())
+    properties.getProperty("API_KEY")
     buildTypes {
         release {
+            buildConfigField("String", "API_KEY", properties.getProperty("API_KEY"))
+            buildConfigField("String", "BASE_URL", properties.getProperty("BASE_URL"))
             isMinifyEnabled = false
             proguardFiles(
                 getDefaultProguardFile("proguard-android-optimize.txt"),
@@ -48,6 +55,8 @@ dependencies {
 
     implementation("io.ktor:ktor-client-core:2.3.13")
     implementation("io.ktor:ktor-client-cio:2.3.13")
+    implementation("io.ktor:ktor-client-content-negotiation:2.3.13")
+
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3")
     implementation("io.ktor:ktor-serialization-kotlinx-json:2.3.2")
     implementation("io.insert-koin:koin-core:3.5.3")
