--- data/src/main/java/com/madrid/data/dataSource/remote/dto/artists/ArtistApiResponse.kt
@@ -1,17 +1,0 @@
-package com.madrid.data.dataSource.remote.dto.artists
-
-
-import com.google.gson.annotations.SerializedName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ArtistApiResponse(
-    @SerializedName("page")
-    val page: Int = 0,
-    @SerializedName("results")
-    val results: List<Result> = listOf(),
-    @SerializedName("total_pages")
-    val totalPages: Int = 0,
-    @SerializedName("total_results")
-    val totalResults: Int = 0
-)
