--- data/src/main/java/com/madrid/data/dataSource/remote/dto/multi/KnownFor.kt
@@ -1,39 +1,0 @@
-package com.madrid.data.dataSource.remote.dto.multi
-
-
-import com.google.gson.annotations.SerializedName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class KnownFor(
-    @SerializedName("adult")
-    val adult: Boolean = false,
-    @SerializedName("backdrop_path")
-    val backdropPath: String = "",
-    @SerializedName("genre_ids")
-    val genreIds: List<Int> = listOf(),
-    @SerializedName("id")
-    val id: Int = 0,
-    @SerializedName("media_type")
-    val mediaType: String = "",
-    @SerializedName("original_language")
-    val originalLanguage: String = "",
-    @SerializedName("original_title")
-    val originalTitle: String = "",
-    @SerializedName("overview")
-    val overview: String = "",
-    @SerializedName("popularity")
-    val popularity: Double = 0.0,
-    @SerializedName("poster_path")
-    val posterPath: String = "",
-    @SerializedName("release_date")
-    val releaseDate: String = "",
-    @SerializedName("title")
-    val title: String = "",
-    @SerializedName("video")
-    val video: Boolean = false,
-    @SerializedName("vote_average")
-    val voteAverage: Double = 0.0,
-    @SerializedName("vote_count")
-    val voteCount: Int = 0
-)
