+++ data/src/main/java/com/madrid/data/dataSource/remote/movies/MovieResponse.kt
@@ -1,8 +1,20 @@
-package com.madrid.data.dataSource.remote.movieResponse
+package com.madrid.data.dataSource.remote.movies
 
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
+
+@Serializable
+data class MovieResponse(
+    @SerializedName("page")
+    val page: Int = 0,
+    @SerializedName("results")
+    val results: List<Result> = listOf(),
+    @SerializedName("total_pages")
+    val totalPages: Int = 0,
+    @SerializedName("total_results")
+    val totalResults: Int = 0,
+)
 
 @Serializable
 data class Result(
@@ -25,7 +37,7 @@ data class Result(
     @SerializedName("poster_path")
     val posterPath: String? = "",
     @SerializedName("release_date")
-    val releaseDate: String?= "",
+    val releaseDate: String? = "",
     @SerializedName("title")
     val title: String? = "",
     @SerializedName("video")
