+++ data/src/main/java/com/madrid/data/dataSource/remote/series/SeriesResponse.kt
@@ -1,11 +1,23 @@
-package com.madrid.data.dataSource.remote.dto.artists
-
+package com.madrid.data.dataSource.remote.series
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class KnownFor(
+data class SeriesResponse(
+    @SerializedName("page")
+    val page: Int = 0,
+    @SerializedName("results")
+    val results: List<Result> = listOf(),
+    @SerializedName("total_pages")
+    val totalPages: Int = 0,
+    @SerializedName("total_results")
+    val totalResults: Int = 0,
+)
+
+
+@Serializable
+data class Result(
     @SerializedName("adult")
     val adult: Boolean = false,
     @SerializedName("backdrop_path")
@@ -14,24 +26,22 @@ data class KnownFor(
     val genreIds: List<Int> = listOf(),
     @SerializedName("id")
     val id: Int = 0,
-    @SerializedName("media_type")
-    val mediaType: String = "",
+    @SerializedName("origin_country")
+    val originCountry: List<String> = listOf(),
     @SerializedName("original_language")
     val originalLanguage: String = "",
-    @SerializedName("original_title")
-    val originalTitle: String = "",
+    @SerializedName("original_name")
+    val originalName: String = "",
     @SerializedName("overview")
     val overview: String = "",
     @SerializedName("popularity")
     val popularity: Double = 0.0,
     @SerializedName("poster_path")
-    val posterPath: String = "",
-    @SerializedName("release_date")
-    val releaseDate: String = "",
-    @SerializedName("title")
-    val title: String = "",
-    @SerializedName("video")
-    val video: Boolean = false,
+    val posterPath: String?,
+    @SerializedName("first_air_date")
+    val firstAirDate: String = "",
+    @SerializedName("name")
+    val name: String = "",
     @SerializedName("vote_average")
     val voteAverage: Double = 0.0,
     @SerializedName("vote_count")
