+++ presentation/src/main/java/com/madrid/presentation/component/screens/searchscreen/SearchListViewModel.kt
@@ -1,45 +1,32 @@
-package com.madrid.presentation.component.screens.searchscreen
+package com.madrid.presentation.component.screens.SearchScreen
 
-
+import androidx.lifecycle.ViewModel
+import com.madrid.data.InMemoryRecentSearchRepository
+import com.madrid.domain.RecentSearchItem
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import androidx.lifecycle.ViewModel
-import com.madrid.domain.RecentSearchRepository
-import com.madrid.domain.RecentSearchItem
-import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.launch
 
-class SearchListViewModel(
-    private val repo: RecentSearchRepository
-) : ViewModel() {
+class SearchListViewModel(private val repo: InMemoryRecentSearchRepository) : ViewModel() {
     private val _recentSearches = MutableStateFlow<List<RecentSearchItem>>(emptyList())
     val recentSearches: StateFlow<List<RecentSearchItem>> = _recentSearches.asStateFlow()
 
     fun loadRecentSearches() {
-        viewModelScope.launch {
-            _recentSearches.value = repo.getRecentSearches()
-        }
+        _recentSearches.value = repo.getRecentSearches()
     }
 
     fun addRecentSearch(item: RecentSearchItem) {
-        viewModelScope.launch {
-            repo.addRecentSearch(item)
-            loadRecentSearches()
-        }
+        repo.addRecentSearch(item)
+        _recentSearches.value = repo.getRecentSearches()
     }
 
-    fun removeRecentSearch(item: RecentSearchItem) {
-        viewModelScope.launch {
-            repo.removeRecentSearch(item.id.toString())
-            loadRecentSearches()
-        }
+    fun removeRecentSearch(id: String) {
+        repo.removeRecentSearch(id)
+        _recentSearches.value = repo.getRecentSearches()
     }
 
-    fun clearAllRecentSearches() {
-        viewModelScope.launch {
-            repo.clearAllRecentSearches()
-            loadRecentSearches()
-        }
+    fun clearAll() {
+        repo.clearAllRecentSearches()
+        _recentSearches.value = repo.getRecentSearches()
     }
 }
