+++ secret.properties
@@ -1,0 +1,2 @@
+API_KEY="go get your api key"
+BASE_URL="https://api.themoviedb.org/3/"
